rag-langgraph/requirements.txt
langchain==0.1.0
langgraph==0.0.26
langchain-openai==0.0.5
langchain-community==0.0.10
chromadb==0.4.22
pypdf==3.17.4
python-dotenv==1.0.0
tiktoken==0.5.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
